extract interfaces and correct types
